Moved general-purpose functions to misc - utils keeps ensemble utils
